+++ seminar_1/task_2/gukasian.vladimir/main.go
@@ -58,67 +58,6 @@ func (list *List) reverseList() {
     list.head = prev
 }
 
-func (list *List) deleteElem(val int) {
-    dummy := new(Node)
-    dummy.next = list.head 
-
-    prev := dummy 
-    cur  := list.head
-
-    for cur != nil {
-        if cur.data == val {
-            prev.next = cur.next 
-            list.size--
-        } else {
-            prev = cur
-        }
-
-        cur = cur.next
-    }
-
-    list.head = dummy.next
-
-}
-
-func (list *List) isRing() bool {
-    if list.head == nil {
-        return false
-    }
-    
-    slow := list.head 
-    fast := list.head.next 
-
-    for slow != nil && fast != nil {
-        if (slow == fast) {
-             return true
-        }
-
-        slow = slow.next 
-        fast = fast.next 
-
-        if (fast != nil) {
-            fast = fast.next
-        }
-    }
-
-    return true 
-}
-
-func (list *List) getNodeByVal(val int) *Node {
-    cur := list.head
-
-    for cur != nil {
-        if cur.data == val {
-          return cur
-         }
-
-        cur = cur.next
-    }
-
-    return nil 
-}
-
-
 func main() {
     capacity := uint(5)
     list     := &List{cap: capacity} 
@@ -142,29 +81,6 @@ func main() {
     list.reverseList()    
     list.Print()
 
-    list.deleteElem(3)
-    list.Print()
-
-    list.deleteElem(5)
-    list.Print()
-
-    list.Append(1)
-    list.deleteElem(1)
-    list.Print()
-    
-    list.Append(1)
-    list.Append(14)
-    
-    cur := list.getNodeByVal(14)
-    cur.next = list.head 
-    
-    list.Print()
-    
-    if list.isRing() {
-        fmt.Println("List is looped")
-    } else {
-        fmt.Println("List is straight")
-    }
 }
