+++ seminar_1/task_4/gukasian.vladimir/main.go
